pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "panscape"
version = "0.1.0"
description = "PanScape: species-level pangenome scaffolding, mapping, and strain deconvolution CLI."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "PanScape contributors" }]
dependencies = [
  "typer>=0.12.5,<1.0",
  "rich>=13.7.1,<14.0",
  "pydantic>=2.8.0,<3.0",
  "PyYAML>=6.0.1,<7.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "black>=24.4.2",
  "ruff>=0.5.6",
  "mypy>=1.10.0",
  "types-PyYAML>=6.0.12.20241230"
]

[project.scripts]
panscape = "panscape.cli:app"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
pretty = true
mypy_path = "src"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
